Refactor type definitions
