+++ platforms/hermes-javascript/src/api/types/Feedback.ts
@@ -1,16 +1,8 @@
+import { NotificationMessage, NotificationMessageLegacy } from './messages'
+
 export namespace FeedbackTypes {
-
-    export type NotificationMessage = {
-        siteId: string,
-        sessionId?: string
-    } | {
-        site_id: string,
-        session_id?: string
-    }
-
     export type publishMessagesList = {
-        notification_on: NotificationMessage,
-        notification_off: NotificationMessage
+        notification_on: NotificationMessage | NotificationMessageLegacy,
+        notification_off: NotificationMessage | NotificationMessageLegacy
     }
-
 }
