+++ platforms/hermes-javascript/src/api/types/enums/grain.ts
@@ -1,0 +1,21 @@
+export enum grain {
+    year = 'Year',
+    quarter = 'Quarter',
+    month = 'Month',
+    week = 'Week',
+    day = 'Day',
+    hour = 'Hour',
+    minute = 'Minute',
+    second = 'Second'
+}
+
+export enum grain_legacy {
+    year = 0,
+    quarter,
+    month,
+    week,
+    day,
+    hour,
+    minute,
+    second
+}
