+++ platforms/hermes-javascript/src/api/types/enums/initType.ts
@@ -1,0 +1,9 @@
+export enum initType {
+    action = 'action',
+    notification = 'notification'
+}
+
+export enum initType_legacy {
+    action = 1,
+    notification
+}
