+++ platforms/hermes-javascript/src/api/types/enums/injectionKind.ts
@@ -1,0 +1,9 @@
+export enum injectionKind {
+    add = 'add',
+    addFromVanilla = 'addFromVanilla'
+}
+
+export enum injectionKind_legacy {
+    add = 1,
+    addFromVanilla
+}
