+++ platforms/hermes-javascript/src/api/types/enums/precision.ts
@@ -1,0 +1,9 @@
+export enum precision {
+    approximate = 'Approximate',
+    exact = 'Exact'
+}
+
+export enum precision_legacy {
+    approximate = 0,
+    exact
+}
