+++ platforms/hermes-javascript/src/api/types/enums/slotType.ts
@@ -1,0 +1,29 @@
+export enum slotType {
+    custom = 'Custom',
+    number = 'Number',
+    ordinal = 'Ordinal',
+    instantTime = 'InstantTime',
+    timeInterval = 'TimeInterval',
+    amountOfMoney = 'AmountOfMoney',
+    temperature = 'Temperature',
+    duration = 'Duration',
+    percentage = 'Percentage',
+    musicAlbum = 'MusicAlbum',
+    musicArtist = 'MusicArtist',
+    musicTrack = 'MusicTrack'
+}
+
+export enum slotType_legacy {
+    custom = 1,
+    number,
+    ordinal,
+    instantTime,
+    timeInterval,
+    amountOfMoney,
+    temperature,
+    duration,
+    percentage,
+    musicAlbum,
+    musicArtist,
+    musicTrack
+}
