+++ platforms/hermes-javascript/src/api/types/enums/terminationType.ts
@@ -1,0 +1,17 @@
+export enum terminationType {
+    nominal = 'nominal',
+    siteUnavailable = 'siteUnavailable',
+    abortedByUser = 'abortedByUser',
+    intentNotRecognized = 'intentNotRecognized',
+    timeout = 'timeout',
+    error = 'error'
+}
+
+export enum terminationType_legacy {
+    nominal = 1,
+    unavailable,
+    abortedByUser,
+    intentNotRecognized,
+    timeout,
+    error
+}
