+++ platforms/hermes-javascript/src/api/types/messages/EndSessionMessage.ts
@@ -1,0 +1,9 @@
+export type EndSessionMessage = {
+    sessionId: string,
+    text?: string
+}
+
+export type EndSessionMessageLegacy = {
+    session_id: string,
+    text?: string
+}
