+++ platforms/hermes-javascript/src/api/types/messages/InjectionStatusMessage.ts
@@ -1,0 +1,7 @@
+export type InjectionStatusMessage = {
+    lastInjectionDate: string
+}
+
+export type InjectionStatusMessageLegacy = {
+    last_injection_date: string
+}
