+++ platforms/hermes-javascript/src/api/types/messages/IntentNotRecognizedMessage.ts
@@ -1,0 +1,15 @@
+export type IntentNotRecognizedMessage = {
+    sessionId: string,
+    siteId: string,
+    input: string,
+    confidence_score: number
+    customData?: string,
+}
+
+export type IntentNotRecognizedMessageLegacy = {
+    session_id: string,
+    site_id: string,
+    input: string,
+    confidence_score: number
+    custom_data?: string,
+}
