+++ platforms/hermes-javascript/src/api/types/messages/NotificationMessage.ts
@@ -1,0 +1,9 @@
+export type NotificationMessage = {
+    siteId: string,
+    sessionId?: string
+}
+
+export type NotificationMessageLegacy = {
+    site_id: string,
+    session_id?: string
+}
