+++ platforms/hermes-javascript/src/api/types/messages/PlayFinishedMessage.ts
@@ -1,0 +1,9 @@
+export type PlayFinishedMessage = {
+    id: string,
+    siteId: string
+}
+
+export type PlayFinishedMessageLegacy = {
+    id: string,
+    site_id: string
+}
