+++ platforms/hermes-javascript/src/api/types/messages/SessionQueuedMessage.ts
@@ -1,0 +1,11 @@
+export type SessionQueuedMessage = {
+    sessionId: string,
+    siteId: string,
+    customData?: string
+}
+
+export type SessionQueuedMessageLegacy = {
+    session_id: string,
+    site_id: string,
+    custom_data?: string
+}
