+++ platforms/hermes-javascript/src/api/types/messages/SessionStartedMessage.ts
@@ -1,0 +1,11 @@
+export type SessionStartedMessage = {
+    sessionId: string,
+    siteId: string
+    customData?: string
+}
+
+export type SessionStartedMessageLegacy =  {
+    session_id: string,
+    site_id: string
+    custom_data?: string
+}
