+++ platforms/hermes-javascript/src/api/types/messages/index.ts
@@ -1,0 +1,13 @@
+export * from './PlayAudioMessage'
+export * from './PlayFinishedMessage'
+export * from './StartSessionMessage'
+export * from './ContinueSessionMessage'
+export * from './EndSessionMessage'
+export * from './IntentMessage'
+export * from './IntentNotRecognizedMessage'
+export * from './SessionEndedMessage'
+export * from './SessionQueuedMessage'
+export * from './SessionStartedMessage'
+export * from './NotificationMessage'
+export * from './InjectionRequestMessage'
+export * from './InjectionStatusMessage'
